requirements.txt
PyQt5
requests
PyPDF2
openpyxl
python-docx
numpy